#50 move map with futures into own file
